pyproject.toml
[project]
name = "portal-web-job-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.1",
    "fastapi>=0.124.4",
    "pymongo>=4.15.5",
    "python-dotenv>=1.2.1",
    "python-jobspy>=1.1.82",
    "schedule>=1.2.2",
    "sentence-transformers>=5.2.0",
]
